+++ tests/BashTest.php
@@ -38,4 +38,22 @@ class BashTest extends RocketeerTests
 		$status = $this->task->checkStatus(null);
 		$this->assertNull($status);
 	}
+
+	public function testCanForgetCredentialsIfInvalid()
+	{
+		$this->app['rocketeer.server']->setValue('credentials', array(
+			'repository' => 'https://Anahkiasen@bitbucket.org/Anahkiasen/registry.git',
+			'username'   => 'Anahkiasen',
+			'password'   => 'baz',
+		));
+
+		// Create fake remote
+		$remote = clone $this->getRemote();
+		$remote->shouldReceive('status')->andReturn(1);
+		$task = $this->task();
+		$task->remote = $remote;
+
+		$task->cloneRepository($this->server.'/test');
+		$this->assertNull($this->app['rocketeer.server']->getValue('credentials'));
+	}
 }