+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 - Ability to select which severs a Task executes on, on a per-task basis
 
+### 0.6.5 (develop)
+
+**- Make Rocketeer prompt for both server and SCM credentials if they're not stored**
+**- `artisan deploy` now deploys the project if the `--version` flat is not passed**
+- Make Rocketeer forget invalid credentials provided by prompt
+- Fix a bug where incorrect SCM urls would be generated
+
 ### 0.6.4 (stable)
 
 - Make the output of commands in realtime when `--verbose` instead of when the command is done